Nerf monk a bit but give armor toughness at lvl 19
Closes #9

Slowed down the rate at which you mine with fists and capped it at
diamond level. Lowered the speed buffs and removed step up as these are
dangerous on a sky island!

To make up for it add half the armor toughness of a full diamond set
at level 19 (but removed potion immunity as its unclear what that even
does)