Change patchouli books to show all content
It is pretty glitchy getting a players achivements to sync properly,
often requiring a player to re-join multiple times before they are able
to read their mod's patchouli book. This was paticularly annoying when
trying to play through natures aura, so hopefully this will allow a
player to get all the info they need more easily to progress through the
mod w/o breaking immersion to go to github to lookup how to do a thing
that the have already unlocked.