Bump mysgear drill speed & removeAndHide dark/redmatter blocks
Closes #66

Removes Dark/Red Matter blocks and furnace as they are a PITA to break
w/o the corresponding pickaxes. The drills can do it but are slow, see
link in script for reasoning.

The embers "Automatic Breaker" seems to break anything almost
immedeately (even when not in creative mode it seemed like...) So the
drills being bumped up orders of magnitude still are very slow compared
to that unpowered breaker... meh