Oh man I added Psi & Magical Psi lmao fun stuff!
This mod doesn't quite fit in with the othes imo as it natively uses a
gun and has a tech armor looking design... however it seems to not
depend on resources / EMC anyway so might be an easy add w/o needing muh
onfig given there is a magia reskin available. give it a go!